Cleaned up few stories and moved few arround
